update ConfigFileLoader in common
