+++ user_sync/config/user_sync.py
@@ -898,172 +898,6 @@ class DictConfig(ObjectConfig):
         return keyring.get_password(service_name=secure_value_key, username=user_name)
 
 
-class ConfigFileLoader:
-
-    def __init__(self, encoding: str, root_path_keys: dict, sub_path_keys: dict):
-        self.encoding = encoding
-        self.root_path_keys = root_path_keys
-        self.sub_path_keys = sub_path_keys
-
-    def load_root_config(self, filename: str) -> dict:
-        """
-        loads the specified file as a root configuration file. This basically
-        means that on top of loading the file as a yaml file into a dictionary,
-        it will apply the ROOT_CONFIG_PATH_KEYS to the dictionary to replace
-        the specified paths with absolute path values that are resolved
-        relative to the given configuration's filename.
-        type filename: str
-        rtype dict
-        """
-        return self.load_from_yaml(filename, self.root_path_keys)
-
-    def load_sub_config(self, filename: str) -> dict:
-        """
-        same as load_root_config, but applies SUB_CONFIG_PATH_KEYS to the
-        dictionary loaded from the yaml file.
-        """
-        return self.load_from_yaml(filename, self.sub_path_keys)
-
-    def load_other_config(self, filename: str) -> dict:
-        """
-        same as load_root_config, but does no post-processing.
-        :type filename: str
-        """
-        return self.load_from_yaml(filename, {})
-
-    def load_from_yaml(self, filepath, path_keys):
-        """
-        loads a yaml file, processes the resulting dict to adapt values for keys
-        (the path to which is defined in path_keys) to a value that represents
-        a file reference relative to the source file being loaded, and returns the
-        processed dict.
-        :param filepath: the file to load yaml from
-        :param path_keys: a dict whose keys are "path_keys" such as /key1/key2/key3
-                          and whose values are tuples: (must_exist, can_have_subdict, default_val)
-                          which are options on the value of the key whose values
-                          are path expanded: must the path exist, can it be a list of paths
-                          that contains sub-dictionaries whose values are paths, and
-                          does the key have a default value so that must be added to
-                          the dictionary if there is not already a value found.
-        """
-        print(filepath)
-        if filepath.startswith('$(') and filepath.endswith(')'):
-            raise AssertionException("Shell execution is no longer supported: {}".format(filepath))
-
-        filepath = os.path.abspath(filepath)
-        if not os.path.isfile(filepath):
-            raise AssertionException('No such configuration file: {}'.format(filepath))
-        filename = os.path.split(filepath)[1]
-        dirpath = os.path.dirname(filepath)
-        try:
-            with open(filepath, 'rb', 1) as input_file:
-                byte_string = input_file.read()
-                yml = yaml.safe_load(byte_string.decode(self.encoding, 'strict'))
-        except IOError as e:
-            # if a file operation error occurred while loading the
-            # configuration file, swallow up the exception and re-raise it
-            # as an configuration loader exception.
-            raise AssertionException("Error reading configuration file '{}': {}".format(filepath, e))
-        except UnicodeDecodeError as e:
-            # as above, but in case of encoding errors
-            raise AssertionException("Encoding error in configuration file '{}: {}".format(filepath, e))
-        except yaml.error.MarkedYAMLError as e:
-            # as above, but in case of parse errors
-            raise AssertionException("Error parsing configuration file '{}': {}".format(filepath, e))
-
-        # process the content of the dict
-        if yml is None:
-            # empty YML files are parsed as None
-            return {}
-        if not isinstance(yml, dict):
-            # malformed YML files produce a non-dictionary
-            raise AssertionException("Configuration file or command '{}' does not contain settings".format(filepath))
-        for path_key, options in six.iteritems(path_keys):
-            key_path = path_key
-            keys = path_key.split('/')
-            self.process_path_key(dirpath, filename, key_path, yml, keys, 1, *options)
-        return yml
-
-    def process_path_key(self, dirpath, filename, key_path, dictionary, keys, level, must_exist, can_have_subdict, default_val):
-        """
-        this function is given the list of keys in the current key_path, and searches
-        recursively into the given dictionary until it finds the designated value, and then
-        resolves relative values in that value to abspaths based on the current filename.
-        If a default value for the key_path is given, and no value is found in the dictionary,
-        then the key_path is added to the dictionary with the expanded default value.
-        type dictionary: dict
-        type keys: list
-        type level: int
-        type must_exist: boolean
-        type can_have_subdict: boolean
-        type default_val: any
-        """
-        # found the key_path, process values
-        if level == len(keys) - 1:
-            key = keys[level]
-            # if a wildcard is specified at this level, that means we
-            # should process all keys as path values
-            if key == "*":
-                for key, val in six.iteritems(dictionary):
-                    dictionary[key] = self.process_path_value(dirpath, filename, key_path, val, must_exist, can_have_subdict)
-            elif key in dictionary:
-                dictionary[key] = self.process_path_value(dirpath, filename, key_path, dictionary[key], must_exist, can_have_subdict)
-            # key was not found, but default value was set, so apply it
-            elif default_val:
-                dictionary[key] = self.relative_path(dirpath, filename, key_path, default_val, must_exist)
-        # otherwise recurse deeper into the dict
-        elif level < len(keys) - 1:
-            key = keys[level]
-            if key in dictionary:
-                # if the key refers to a dictionary, recurse into it to go
-                # further down the path key
-                if isinstance(dictionary[key], dict):
-                    self.process_path_key(dirpath, filename, key_path, dictionary[key], keys, level + 1,
-                                          must_exist, can_have_subdict, default_val)
-            # if the key was not found, but a default value is specified,
-            # drill down further to set the default value
-            elif default_val:
-                dictionary[key] = {}
-                self.process_path_key(dirpath, filename, key_path, dictionary[key], keys, level + 1,
-                                      must_exist, can_have_subdict, default_val)
-
-    def process_path_value(self, dirpath, filename, key_path, val, must_exist, can_have_subdict):
-        """
-        does the relative path processing for a value from the dictionary,
-        which can be a string, a list of strings, or a list of strings
-        and "tagged" strings (sub-dictionaries whose values are strings)
-        :param val: the value we are processing, for error messages
-        :param must_exist: whether there must be a value
-        :param can_have_subdict: whether the value can be a tagged string
-        """
-        if isinstance(val, six.string_types):
-            return self.relative_path(dirpath, filename, key_path, val, must_exist)
-        elif isinstance(val, list):
-            vals = []
-            for entry in val:
-                if can_have_subdict and isinstance(entry, dict):
-                    for subkey, subval in six.iteritems(entry):
-                        vals.append({subkey: self.relative_path(dirpath, filename, key_path, subval, must_exist)})
-                else:
-                    vals.append(self.relative_path(dirpath, filename, key_path, entry, must_exist))
-            return vals
-
-    @staticmethod
-    def relative_path(dirpath, filename, key_path, val, must_exist):
-        """
-        returns an absolute path that is resolved relative to the file being loaded
-        """
-        if not isinstance(val, six.string_types):
-            raise AssertionException("Expected pathname for setting {} in config file {}".format(key_path, filename))
-        if val.startswith('$(') and val.endswith(')'):
-            raise AssertionException("Shell execution is no longer supported: {}".format(val))
-        if dirpath and not os.path.isabs(val):
-            val = os.path.abspath(os.path.join(dirpath, val))
-        if must_exist and not os.path.isfile(val):
-            raise AssertionException('In setting {} in config file {}: No such file {}'.format(key_path, filename, val))
-        return val
-
-
 class OptionsBuilder(object):
     def __init__(self, default_config):
         """
